Define type Bitfield.t for the bitfield strings in the protocol.
Bitfield.t = string and is just a type to identify those strings that
represent the protocol bitfield strings.

Besides, we now use (tentatively) Int.Set.t in application layer to manipulate
the pieces owned by the peers (in Peer) and not requested (in File).
This make the application layer more abstract.

The next step is to make Bitfield.t private and decide if we should stick to
Bitset.t for representing sets or transition everything to Int.Set.t
(performance vs code readability).