Split module File in two. File and PFile.
Previously, we only dealt with "single file torrent" but the code was getting
messy with persistence for multiple file.

Now File deals with the whole file to be downloaded from the peers
(ie. essentially the array of pieces).
PFile deals with persistence. Every file to be written on disk is a PFile.
That includes the bitset, and all the files that make up the "whole file".

Right now, File.create and File.close call the (not yet implemented)
PFile function.