Completed Pfile module.
This includes:
 - assigning Pfiles to pieces by aligned pfiles to pieces
 - writing the read/write function in PFile
 - calling them at application startup/closing time

Still some issues with concurrent write/lseek. Now there is a PFile.debug
file that mocks Reader.read and Writer.write function so we can still
run the application.

Besides this, this is functional and documented but the design will have to be
reworked eventually.

Also added a new (legal) multi-torrent file for testing as well as the
makefile directives to run it.