Moved Extract_bencode to Torrent
Extract_bencode was used to extract bencode from torrent files, and
from the tracker replies.

Now Torrent.t is the structure that contains the torrent info, and
bencode extraction of tracker replies is done in Tracker_client. There
are a couple of small function that are duplicated but it's cleaner this way.