Set up a pipe of writing jobs between App_layer and Pers.
When new pieces are downloaded, infos (essentially piece index) are sent via a
pipe to a Pers master thread that write the pieces at the right place
on the disk.

Pretty much all the code is there, just need to make it work properly.
Once this works, still need to iterate again on the design.