moved file / pers / peers handshake in Start (from pwp).
This can generate errors that are easier to deal with in Start
(e.g. opening files).

If the client can serve multiple torrents (not yet the case), handshake
needs to be done in order to know which file to transfer (and add the peer
to Pwp which is torrent specific).

This alleviate module Pwp that only deals with the protocol and doesn't need
to bother initalizing these structures and propagating errors.