Reworked File/Pieces interfaces and separation.
Piece.status becomes File.pieces_status so all the info related to the status of
a piece is now in File. It provides better encapulsation (File.bitset and
Pieces.status contained partly the same information and had to be updated
together).  Actually, we could even get rid of bitset since we can re-generate it
from File.pieces_status. The only reason to keep it around is for efficiency
(but right now doesn't change anything since we have a dumb implementation
of bitset).