Changed how pieces are requested.
This was previously done using polling to see if new pieces were available.
We do it now based on events that may change the availability of pieces
(e.g. Have messages, Unchoke).

We sort the pieces "rarest" first and pick the first few ones that
leave us under a fixed number of pending requests (arbitrarly set to 10).

We recompute this all the time which is very inefficient. It'd be better
to use an appropriate datastructure that would be updated at each new event.

Besides that, we still don't serve any requests, even though it's implemented.
Maybe a bad choice of torrent file.

Right now, we suppose at the beginning that we are non-choking and interested
for all peers but this should be the opposite and be dynamically changed.