simplified module Node
just keep information relevant to a connection.
everything else (status, id) is kept in Krpc.

In the previous version, we assumed we always got an open socket to all nodes
in the table. Now we create the socket whenever we communicate with the peer
and we only keep an address in the table.