Split Peer into Peer/State
All the state relative to the protocol is now in module State.
Peer is only an interface to talk to a peer.
Now in Pwp, we use the "peer_info" structure that combines both info.

This is done for several reasons.
- It will be easier to have efficient datastructure to represent the state
  of the swarm, instead of having it split in different values.
- Now that we are implementing DHT, we end up with peers that don't know
  anything about the meta-info. Instead of creating them and changing their
  state later on when we know more (which we already did for the peers we got
  from the server), it's easier to create them once for all and keep the state
  in Pwp.
- It keeps the modules size  more tractable

Tests seem to be pass, but there's more cleaning up to do.