optional meta field in pwp.
a subset of fields in pwp now becomes optional. This is because we want
to start adding peers before we know the meta info. In that case, we
can't initialize the whole structure (e.g. network file) because we don't
have enough information.

This makes things a little more complicated but I haven't found a better way.