Setting up things to deal with magnets
There is now an homogeneous treatment of magnet and torrent. In Start
we create Pwp.t the same way for both input. Then we launch one or two
threads to lookup for peers.
 - for magnet, we only use DHT
 - for torrent, we use DHT and tracker

Then, in the case of torrent, we give Pwp the meta-info that allows it to
lookup create the Network_file.

If the network_file is not available, peers block on an ivar.

Next things to do is that instead of blocking, we'll tell the peers to
lookup for the meta-info via the DHT.